requirements.txt
## The following requirements were added by pip freeze:
ruamel.yaml==0.17.2
ruamel.yaml.clib==0.2.2
